fix: fixed pipes, and ongoing functions

+++ x64BareBones-master/Kernel/include/pipes.h
@@ -9,7 +9,7 @@
 #define MAX_PIPES 20
 #define PIPE_BUFFER_SIZE 128
 
-int64_t createPipe(int pipeID, int fd[2]);
+int64_t createPipe(int pipeID, int16_t fd[2]);
 void destroyPipe(int pipeID);
 uint64_t writePipe(int pipeID, const char * source);
 uint64_t readPipe(int pipeID, char * destination);
